+++ app/controllers/datasets_controller.rb
@@ -1,10 +1,10 @@
 class DatasetsController < ApplicationController
   before_action :authenticate_user_from_token!, only: [ :json_manifest, :manifest, :files, :upload_graph, :refresh_dictionary, :upload_dataset_csv, :editor ]
-  before_action :authenticate_user!,        only: [ :new, :edit, :create, :update, :destroy, :audits, :requests, :request_access, :set_access, :create_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync, :set_public_file, :reset_index ]
+  before_action :authenticate_user!,        only: [ :new, :edit, :create, :update, :destroy, :audits, :requests, :create_access, :remove_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync, :set_public_file, :reset_index ]
   before_action :check_system_admin,        only: [ :new, :create, :destroy, :pull_changes, :sync ]
-  before_action :set_viewable_dataset,      only: [ :show, :json_manifest, :manifest, :logo, :images, :files, :access, :pages, :request_access, :search, :editor ]
-  before_action :set_editable_dataset,      only: [ :edit, :update, :destroy, :audits, :requests, :set_access, :create_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync, :set_public_file, :reset_index, :upload_graph, :refresh_dictionary, :upload_dataset_csv ]
-  before_action :redirect_without_dataset,  only: [ :show, :json_manifest, :manifest, :logo, :images, :files, :access, :pages, :request_access, :set_access, :create_access, :search, :edit, :update, :destroy, :audits, :requests, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync, :set_public_file, :reset_index, :upload_graph, :refresh_dictionary, :upload_dataset_csv, :editor ]
+  before_action :set_viewable_dataset,      only: [ :show, :json_manifest, :manifest, :logo, :images, :files, :access, :pages, :search, :editor ]
+  before_action :set_editable_dataset,      only: [ :edit, :update, :destroy, :audits, :requests, :create_access, :remove_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync, :set_public_file, :reset_index, :upload_graph, :refresh_dictionary, :upload_dataset_csv ]
+  before_action :redirect_without_dataset,  only: [ :show, :json_manifest, :manifest, :logo, :images, :files, :access, :pages, :create_access, :remove_access, :search, :edit, :update, :destroy, :audits, :requests, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync, :set_public_file, :reset_index, :upload_graph, :refresh_dictionary, :upload_dataset_csv, :editor ]
 
   skip_before_action :verify_authenticity_token, only: [ :upload_graph, :upload_dataset_csv ]
 
@@ -111,36 +111,16 @@ class DatasetsController < ApplicationController
     redirect_to files_dataset_path(@dataset, path: @dataset.find_file_folder(params[:path]))
   end
 
-  def request_access
-    if @dataset_user = @dataset.dataset_users.where( user_id: current_user.id ).first
-      # Dataset access has already been requested
-    else
-      @dataset_user = @dataset.dataset_users.create( user_id: current_user.id, editor: false, approved: nil )
-      @dataset.editors.each do |editor|
-        UserMailer.dataset_access_requested(@dataset_user, editor).deliver_later if Rails.env.production?
-      end
-    end
-    if params[:path]
-      redirect_to files_dataset_path(@dataset, path: params[:path])
-    else
-      redirect_to submissions_path
-    end
-  end
-
-  def set_access
-    if @dataset_user = @dataset.dataset_users.find_by_id(params[:dataset_user_id])
-      @dataset_user.update( editor: params[:editor], approved: params[:approved] )
-      if @dataset_user.approved? and not @dataset_user.email_sent?
-        UserMailer.dataset_access_approved(@dataset_user, current_user).deliver_later if Rails.env.production?
-        @dataset_user.update email_sent: true
-      end
-    end
-    redirect_to requests_dataset_path(@dataset, dataset_user_id: @dataset_user ? @dataset_user.id : nil)
-  end
-
   def create_access
-    @dataset_user = @dataset.dataset_users.where( user_id: params[:user_id] ).first_or_create
+    @dataset_user = @dataset.dataset_users.where( user_id: params[:user_id], role: params[:role] ).first_or_create
     redirect_to requests_dataset_path(@dataset, dataset_user_id: @dataset_user ? @dataset_user.id : nil)
+  end
+
+  def remove_access
+    if @dataset_user = @dataset.dataset_users.find_by_id(params[:dataset_user_id])
+      @dataset_user.destroy
+    end
+    redirect_to requests_dataset_path(@dataset)
   end
 
   # GET /datasets/1/file_audits