+++ app/models/dataset.rb
@@ -9,10 +9,9 @@ class Dataset < ActiveRecord::Base
 
   # Named Scopes
   scope :release_scheduled, -> { current.where( public: true ).where.not( release_date: nil )}
-  scope :with_editor, lambda { |arg| where('datasets.user_id IN (?) or datasets.id in (select dataset_users.dataset_id from dataset_users where dataset_users.user_id = ? and dataset_users.editor = ? and dataset_users.approved = ?)', arg, arg, true, true ).references(:dataset_users) }
-  # scope :with_viewer, lambda { |arg| where('datasets.user_id IN (?) or datasets.public = ? or datasets.id in (select dataset_users.dataset_id from dataset_users where dataset_users.user_id = ? and dataset_users.approved = ?)', arg, true, arg, true ).references(:dataset_users) }
-  # scope :with_reviewer, lambda { |arg| where('datasets.user_id IN (?) or datasets.id in (select dataset_users.dataset_id from dataset_users where dataset_users.user_id = ? and dataset_users.role = ? and dataset_users.approved = ?)', arg, arg, 'reviewer', true ).references(:dataset_users) }
-  scope :with_viewer, lambda { |arg| where('datasets.user_id IN (?) or datasets.public = ? or datasets.id in (select requests.dataset_id from requests, agreements where requests.agreement_id = agreements.id and agreements.status = ? and agreements.deleted = ? and (agreements.expiration_date IS NULL or agreements.expiration_date >= ?) and agreements.user_id IN (?) )', arg, true, 'approved', false, Date.today, arg ).references(:requests) }
+  scope :with_editor,   lambda { |arg| where(                       'datasets.user_id IN (?) or datasets.id in (select dataset_users.dataset_id from dataset_users where dataset_users.user_id = ? and dataset_users.role = ?)',       arg, arg, 'editor'   ).references(:dataset_users) }
+  scope :with_reviewer, lambda { |arg| where(                                                  'datasets.id in (select dataset_users.dataset_id from dataset_users where dataset_users.user_id = ? and dataset_users.role = ?)',            arg, 'reviewer' ).references(:dataset_users) }
+  scope :with_viewer,   lambda { |arg| where('datasets.public = ? or datasets.user_id IN (?) or datasets.id in (select dataset_users.dataset_id from dataset_users where dataset_users.user_id = ? and dataset_users.role = ?)', true, arg, arg, 'viewer'   ).references(:dataset_users) }
 
   # Model Validation
   validates_presence_of :name, :slug, :user_id
@@ -44,11 +43,11 @@ class Dataset < ActiveRecord::Base
   end
 
   def viewers
-    User.where( id: [self.user_id] + self.dataset_users.where( approved: true ).pluck(:user_id) )
+    User.where( id: [self.user_id] + self.dataset_users.where( role: 'viewer' ).pluck(:user_id) )
   end
 
   def editors
-    User.where( id: [self.user_id] + self.dataset_users.where( approved: true, editor: true ).pluck(:user_id) )
+    User.where( id: [self.user_id] + self.dataset_users.where( role: 'editor' ).pluck(:user_id) )
   end
 
   # def grants_file_access_to?(current_user)