+++ app/models/dataset_user.rb
@@ -1,5 +1,10 @@
 class DatasetUser < ActiveRecord::Base
 
+  ROLES = [['Editor', 'editor'],['Viewer', 'viewer'],['Reviewer', 'reviewer']]
+
+  # Model Validation
+  validates_presence_of :user_id, :role
+
   # Model Relationships
   belongs_to :dataset
   belongs_to :user