+++ db/migrate/20141202185444_add_role_to_dataset_users.rb
@@ -1,0 +1,5 @@
+class AddRoleToDatasetUsers < ActiveRecord::Migration
+  def change
+    add_column :dataset_users, :role, :string
+  end
+end